pyproject.toml
[tool.poetry]
name = "probable-broccoli"
version = "0.1.0"
description = ""
authors = ["Jacek Palczewski <jpalczewski@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fbchat = "^1.9.7"
python-dotenv = "^0.14.0"
IPython = "^7.18.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
